Updated Clippy corrections and format
