+++ .vscode/settings.json
@@ -1,6 +1,5 @@
 {
     "rust-analyzer.linkedProjects": [
         "./Cargo.toml",
-        "./Cargo.toml"
     ]
 }
