+++ src/systems.rs
@@ -7,13 +7,10 @@ pub fn transition_to_gamestate(
     keyboard_input: Res<ButtonInput<KeyCode>>,
     simulation_state: Res<State<AppState>>,
 ) {
-    if keyboard_input.just_pressed(KeyCode::KeyG) {
-        if simulation_state.get() != &AppState::Game {
-            commands.insert_resource(NextState(Some(AppState::Game)));
+    if keyboard_input.just_pressed(KeyCode::KeyG) && simulation_state.get() != &AppState::Game {
+        commands.insert_resource(NextState(Some(AppState::Game)));
         println!("Entered Game State")
-        } 
     }
-        
 }
 
 pub fn transition_to_main_menu_state(
@@ -21,10 +18,8 @@ pub fn transition_to_main_menu_state(
     keyboard_input: Res<ButtonInput<KeyCode>>,
     simulation_state: Res<State<AppState>>,
 ) {
-    if keyboard_input.just_pressed(KeyCode::KeyM) {
-        if simulation_state.get() != &AppState::MainMenu {
-            commands.insert_resource(NextState(Some(AppState::MainMenu)));
+    if keyboard_input.just_pressed(KeyCode::KeyM) && simulation_state.get() != &AppState::MainMenu {
+        commands.insert_resource(NextState(Some(AppState::MainMenu)));
         println!("Entered Main Menu State")
-        }
-    } 
+    }
 }
